requirements.txt
# Python Requirements file.

paramiko
python-dotenv